Move classes to separate files, allow changing colors
